add support for 14 factorial

+++ Asynchronous Programming in Java/src/main/java/org/example/asynchronous/common/FactorialCalculator.java
@@ -2,5 +2,5 @@ package org.example.asynchronous.common;
 
 public interface FactorialCalculator {
 
-  int calculate(int number);
+  long calculate(int number);
 }
